+++ config/Seeds/AdminsSeed.php
@@ -1,15 +1,22 @@
 <?php
+
 declare(strict_types=1);
 
-use Migrations\AbstractSeed;
+use Migrations\BaseSeed;
 use Authentication\PasswordHasher\DefaultPasswordHasher;
 
-class AdminSeed extends AbstractSeed
+/**
+ * Admins seed.
+ */
+class AdminsSeed extends BaseSeed
 {
     /**
      * Run Method.
      *
      * Write your database seeder using this method.
+     *
+     * More information on writing seeds is available here:
+     * https://book.cakephp.org/migrations/4/en/seeding.html
      *
      * @return void
      */